msm: ep_pcie: Handle PERST Deassert from ep_pcie prior to MHI init
In the current implementation, PERST deassert event is handled by
ep_pcie driver only when enumeration is not completed at least once. It
is notified to MHI if initial enumeration is completed, irrespective of
whether the event is registered by MHI or not. With this implementation,
seeing a link down issue in below scenario with warm reset.

Sequence of events:
1. Warm reset is done
2. Device booted to HLOS and enumeration is completed
3. As IPA DMA provide OPS call is not received, device transitioned to
D3cold state
4. Before receiving the above IPA call, PERST deassert event is received
due to a client activity on host side
5. MHI initialization is not completed yet and the above event is not
registered by MHI
6. Device fails to notify the event to MHI as the event is not
registered yet
7. Not handling PERST deassert is leading to link down on host side

To handle this scenario added a logic to notify PERST deassert event to
MHI only if the event is registered by MHI or else handling it from
ep_pcie driver. Similar logic is already present for PERST assert event,
just updated the check condition here, as notify function returns bool
variable now with this change.

Change-Id: I15eac73244588e2fa9d4a1821fe019f5f32bf5e8
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>