serial: msm_geni_serial: capture CTS counter value during data TX
Add support to capture CTS interrupt counter value during the
process of sending data from MSM to BT Soc. This info can be
used to check that the MSM CTS line is not toggling
during the process of preparing data for transmission till
the time the data is on the lines.

Change-Id: I0c772846a53ff0ce3c0fabf104c8c54107408d66
Signed-off-by: Prasanna S <quic_prass@quicinc.com>