arm-smmu: change supend retry option in prepare callback
In case of pm runtime suspend is in progress, then kgsl smmu
system suspend would go for retry by returing -EINPROGRESS.
but this is aborting the complete suspend for a client,
which doesn't want to change their user space to handle reattempt.

to fix this by calling pm_runtime_put_sync_suspend in pm_prepare,
we ensure that rpm_suspend had been executed before suspend.
since call is synchronous, frozen workqueue is not used
and does not prevent execution. changed this return value to -EAGAIN
to retry the prepare again to make pm runtime suspended correctly.

Change-Id: Id602051aaebcc3ddbd16a3e496e8b7f512679f3a
Signed-off-by: Madhu Ananthula <quic_mananthu@quicinc.com>