drivers: goodix_9916: Ensure that coor mode is enabled
M11 touchscreen is configured to run in rawdata mode for further
processing of the touch capacitance data using the touch_report
service.

For making the touchscreen to report the touch coordinates instead
of the raw data, we have to make sure that the rawdata mode is off
(0x90 command), and that coor mode is on (0x91 command).

To achieve this, run brld_set_coor_mode whenever the touch IC resets,
as that makes the touchscreen to run again in rawdata mode, and also
whenever the touch data checksum mismatches.

Change-Id: I562ddc98e93f6d6465c8a2be8c8959c224eb2a9d
Signed-off-by: danielml <daniel@danielml.dev>