drivers: goodix_9916: Avoid repeatedly enabling coor mode
We only want to enable coor mode after the touch IC has been reset
or powered on.

Instead of enabling it each time that an event checksum mismatches
(which can happen even with coor mode enabled), enable it only one time
after the touch driver is probed, and on brl_reset which is already
implemented.

This avoids incorrect touch reports on situations where the checksum
mismatches.

Change-Id: I224f246b4b31258ec1e99e247bfe087b717a755e
Signed-off-by: danielml <daniel@danielml.dev>